Renaming existing namespace MessageBird.Objects.VoiceCalls to MessageBird.Objects.Voice and MessageBird.Resources.VoiceCalls to MessageBird.Resources.Voice so that we can put all other Voice API related objects and resources into these namespaces together
