+++ MessageBird/MessageBird.csproj
@@ -15,4 +15,7 @@
   <ItemGroup>
     <PackageReference Include="Newtonsoft.Json" Version="9.0.1" />
   </ItemGroup>
+  <ItemGroup>
+    <Folder Include="Objects\Common\" />
+  </ItemGroup>
 </Project>
