+++ MessageBird/Resources/Voice/VoiceCallFlowsBaseLists.cs
@@ -1,15 +1,13 @@
 ﻿using System.Text;
-using MessageBird.Objects;
-using MessageBird.Objects.VoiceCalls;
+using MessageBird.Objects.Voice;
 
-namespace MessageBird.Resources.VoiceCalls
+namespace MessageBird.Resources.Voice
 {
-    public class VoiceCallFlowsBaseLists<T> : VoiceCallFlowsResource
+    public class VoiceCallFlowsBaseLists : VoiceCallFlowsResource
     {
         public VoiceCallFlowsBaseLists(string name, VoiceCallFlowList attachedObject)
             : base(name, attachedObject)
         {
-            //
         }
 
         public override string QueryString
