--- MessageBird/Resources/VoiceCalls/VoiceCallFlowLists.cs
@@ -1,9 +1,0 @@
-﻿namespace MessageBird.Resources.VoiceCalls
-{
-    public class VoiceCallFlowLists : VoiceCallFlowsBaseLists<Objects.VoiceCalls.VoiceCallFlow>
-    {
-        public VoiceCallFlowLists()
-            : base("call-flows", new Objects.VoiceCalls.VoiceCallFlowList())
-        { }
-    }
-}
